ref(metrics): move metrics api data endpoint to separate file
