+++ src/sentry/api/endpoints/organization_metrics.py
@@ -13,7 +13,6 @@ from sentry.api.bases.organization import (
     OrganizationPermission,
 )
 from sentry.api.exceptions import ResourceDoesNotExist
-from sentry.api.paginator import GenericOffsetPaginator
 from sentry.api.utils import get_date_range_from_params
 from sentry.auth.elevated_mode import has_elevated_mode
 from sentry.exceptions import InvalidParams
@@ -28,11 +27,9 @@ from sentry.sentry_metrics.use_case_id_registry import (
     get_use_case_id_api_access,
 )
 from sentry.sentry_metrics.utils import string_to_use_case_id
-from sentry.snuba.metrics import QueryDefinition, get_all_tags, get_series, get_single_metric_info
-from sentry.snuba.metrics.utils import DerivedMetricException, DerivedMetricParseException
+from sentry.snuba.metrics import get_all_tags, get_single_metric_info
+from sentry.snuba.metrics.utils import DerivedMetricParseException
 from sentry.snuba.sessions_v2 import InvalidField
-from sentry.types.ratelimit import RateLimit, RateLimitCategory
-from sentry.utils.cursors import Cursor, CursorResult
 
 
 def can_access_use_case_id(request: Request, use_case_id: UseCaseID) -> bool:
@@ -255,84 +252,3 @@ class OrganizationMetricsTagDetailsEndpoint(OrganizationEndpoint):
             return f"No data found for metric: {metric_names[0]} and tag: {tag_name}"
         else:
             return f"No data found for tag: {tag_name}"
-
-
-@region_silo_endpoint
-class OrganizationMetricsDataEndpoint(OrganizationEndpoint):
-    publish_status = {
-        "GET": ApiPublishStatus.EXPERIMENTAL,
-    }
-    owner = ApiOwner.TELEMETRY_EXPERIENCE
-    permission_classes = (OrganizationAndStaffPermission,)
-
-    """Get the time series data for one or more metrics.
-
-    The data can be filtered and grouped by tags.
-    Based on `OrganizationSessionsEndpoint`.
-    """
-
-    # 60 req/s to allow for metric dashboard loading
-    default_rate_limit = RateLimit(limit=60, window=1)
-
-    rate_limits = {
-        "GET": {
-            RateLimitCategory.IP: default_rate_limit,
-            RateLimitCategory.USER: default_rate_limit,
-            RateLimitCategory.ORGANIZATION: default_rate_limit,
-        },
-    }
-
-    # Number of groups returned for each page (applies to old endpoint).
-    default_per_page = 50
-
-    def get(self, request: Request, organization) -> Response:
-        projects = self.get_projects(request, organization)
-
-        def data_fn(offset: int, limit: int):
-            try:
-                query = QueryDefinition(
-                    projects,
-                    request.GET,
-                    allow_mri=True,
-                    paginator_kwargs={"limit": limit, "offset": offset},
-                )
-                data = get_series(
-                    projects,
-                    metrics_query=query.to_metrics_query(),
-                    use_case_id=get_use_case_id(request),
-                    tenant_ids={"organization_id": organization.id},
-                )
-                data["query"] = query.query
-            except (
-                InvalidParams,
-                DerivedMetricException,
-            ) as exc:
-                raise (ParseError(detail=str(exc)))
-            return data
-
-        return self.paginate(
-            request,
-            paginator=MetricsDataSeriesPaginator(data_fn=data_fn),
-            default_per_page=self.default_per_page,
-            max_per_page=100,
-        )
-
-
-class MetricsDataSeriesPaginator(GenericOffsetPaginator):
-    def get_result(self, limit, cursor=None):
-        assert limit > 0
-        offset = cursor.offset if cursor is not None else 0
-        data = self.data_fn(offset=offset, limit=limit + 1)
-
-        if isinstance(data.get("groups"), list):
-            has_more = len(data["groups"]) == limit + 1
-            if has_more:
-                data["groups"].pop()
-        else:
-            raise NotImplementedError
-
-        return CursorResult(
-            data,
-            prev=Cursor(0, max(0, offset - limit), True, offset > 0),
-            next=Cursor(0, max(0, offset + limit), False, has_more),
-        )
